LayoutTests/imported/w3c/web-platform-tests/css/css-text/text-autospace/text-autospace-mixed-001-expected.txt

国国אבג国国
国国אבג国国
国国مُرَبَّعْ国国

FAIL text-autospace-mixed-001 assert_array_equals: expected property 1 to be 45 but got 40 (expected array [40, 45, 72, -47, 87, -58, 103] got [40, 40, 72, -47, 87, -63, 103])
FAIL text-autospace-mixed-001 1 assert_array_equals: expected property 1 to be 45 but got 40 (expected array [40, 45, 24.00390625, 87.99609375, -63, 28, 80] got [40, 40, 24.00390625, 87.99609375, -63, 23, 80])
FAIL text-autospace-mixed-001 2 assert_array_equals: expected property 1 to be 45 but got 40 (expected array [40, 45, 96.015625, -71.015625, -25, 137, -100, -13, 25, 24, -25, 29, 104.015625] got [40, 40, 96.015625, -71.015625, -25, 137, -100, -13, 25, 24, -25, 24, 104.015625])
